Day3/example/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
djangorestframework==3.14.0
psycopg==3.1.8
psycopg-binary==3.1.8
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
